move query tests
