+++ tests/test_queries/test_where.py
@@ -1,107 +1,9 @@
 from __future__ import annotations
 
-import sys
-from collections.abc import Callable, Generator
+from collections.abc import Callable
 from typing import Any, Literal
 
 import pytest
-from graphql import GraphQLSchema, graphql_sync
-from graphql_sqlalchemy.schema import build_schema
-from sqlalchemy import Column, Engine, ForeignKey, Table
-from sqlalchemy.orm import DeclarativeBase, Mapped, Session, mapped_column, registry, relationship
-
-if sys.version_info < (3, 11):
-    from exceptiongroup import ExceptionGroup
-
-
-class Base(DeclarativeBase):
-    registry = registry()
-
-
-article_tag_association = Table(
-    "article_tag",
-    Base.metadata,
-    Column("article_id", ForeignKey("article.id"), primary_key=True),
-    Column("tag_id", ForeignKey("tag.id"), primary_key=True),
-)
-
-
-class Author(Base):
-    __tablename__ = "author"
-    id: Mapped[int] = mapped_column(primary_key=True)
-    name: Mapped[str]
-    articles: Mapped[list[Article]] = relationship(back_populates="author")
-
-
-class Article(Base):
-    __tablename__ = "article"
-    id: Mapped[int] = mapped_column(primary_key=True)
-    title: Mapped[str]
-    author_id: Mapped[int] = mapped_column(ForeignKey("author.id"))
-    author: Mapped[Author] = relationship(back_populates="articles")
-    rating: Mapped[int]
-    tags: Mapped[list[Tag]] = relationship(back_populates="articles", secondary=article_tag_association)
-
-
-class Tag(Base):
-    __tablename__ = "tag"
-    id: Mapped[int] = mapped_column(primary_key=True)
-    name: Mapped[str]
-    articles: Mapped[list[Article]] = relationship(back_populates="tags", secondary=article_tag_association)
-
-
-@pytest.fixture(scope="session")
-def gql_schema() -> GraphQLSchema:
-    return build_schema(Base)
-
-
-@pytest.fixture()
-def example_session(db_engine: Engine, db_session: Session) -> Generator[Session, None, None]:
-    Base.metadata.create_all(bind=db_engine)
-    with db_session.begin():
-        db_session.add(tag_politics := Tag(name="Politics"))
-        db_session.add(tag_sports := Tag(name="Sports"))
-
-        db_session.add(felicias := Author(name="Felicitas"))
-        db_session.add_all(
-            [
-                Article(title="Felicitas good", author=felicias, rating=4),
-                Article(title="Felicitas better", author=felicias, rating=5, tags=[tag_politics, tag_sports]),
-            ]
-        )
-        db_session.add(bjork := Author(name="Bjørk"))
-        db_session.add_all(
-            [
-                Article(title="Bjørk bad", author=bjork, rating=2),
-                Article(title="Bjørk good", author=bjork, rating=4, tags=[tag_politics]),
-            ]
-        )
-        db_session.add(lundth := Author(name="Lundth"))
-        db_session.add_all(
-            [
-                Article(title="Lundth bad", author=lundth, rating=1, tags=[tag_sports]),
-            ]
-        )
-        db_session.commit()
-
-    yield db_session
-
-    Base.metadata.drop_all(bind=db_engine)
-
-
-@pytest.fixture()
-def query_example(example_session: Session, gql_schema: GraphQLSchema) -> Callable[[str], Any]:
-    def query(q: str) -> Any:
-        source = f"query {{ {q} }}"
-        result = graphql_sync(gql_schema, source, context_value={"session": example_session})
-        if example_session._transaction:
-            # TODO: make unnecessary
-            example_session._transaction.close()
-        if result.errors:
-            raise result.errors[0] if len(result.errors) == 1 else ExceptionGroup("Invalid Query", result.errors)
-        return result.data
-
-    return query
 
 
 @pytest.mark.parametrize("filt", ["", "(where: { })"])
@@ -164,9 +66,7 @@ def test_and_or(query_example: Callable[[str], Any], op: Literal["and", "or"], e
         pytest.param("", id="artcl_all"),
     ],
 )
-def test_nested_filter_one2many(
-    db_session: Session, query_example: Callable[[str], Any], filter_author: str, filter_article: str
-) -> None:
+def test_nested_filter_one2many(query_example: Callable[[str], Any], filter_author: str, filter_article: str) -> None:
     data = query_example(
         f"""
         author{filter_author} {{
@@ -185,8 +85,7 @@ def test_nested_filter_one2many(
 
     articles = [article for author in data["author"] for article in author["articles"]]
     article_titles = {article["title"] for article in articles}
-    with db_session.begin():
-        all_article_titles = {row[0] for row in db_session.query(Article.title).all()}
+    all_article_titles = {article["title"] for article in query_example("article { title }")["article"]}
     if filter_article:
         assert article_titles == {"Felicitas good", "Felicitas better", "Bjørk good"}
     elif filter_author:
@@ -198,7 +97,7 @@ def test_nested_filter_one2many(
 def test_nested_filter_many2one(query_example: Callable[[str], Any]) -> None:
     data = query_example(
         """
-        article(where: { author: { name: { _eq: "Lundth" } } }) {
+        article(where: { author: { name: { _in: ["Lundth"] } } }) {
             id title rating
         }
         """
